Gallery now working
