+++ app/src/main/java/hypr/a255bits/com/hypr/CameraFragment/CameraActivity.kt
@@ -1,5 +1,8 @@
 package hypr.a255bits.com.hypr.CameraFragment
 
+import android.app.Activity
+import android.content.Intent
+import android.net.Uri
 import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
 import android.support.v4.app.NavUtils
@@ -13,10 +16,10 @@ import org.jetbrains.anko.clearTop
 import org.jetbrains.anko.intentFor
 
 class CameraActivity : AppCompatActivity(), CameraMVP.view {
-
-    val presenter: CameraPresenter by lazy { CameraPresenter(this) }
-
+    val presenter: CameraPresenter by lazy { CameraPresenter(this, applicationContext) }
+    private val  RESULT_GET_IMAGE: Int = 1
     var indexInJson: Int? = null
+    val galleryFileLocation: Uri = android.provider.MediaStore.Images.Media.EXTERNAL_CONTENT_URI
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -30,10 +33,20 @@ class CameraActivity : AppCompatActivity(), CameraMVP.view {
                 presenter.sendPictureToModel(jpeg)
             }
         })
-        takePicture.setOnClickListener {
-            cameraView.captureImage()
-        }
+        takePicture.setOnClickListener { cameraView.captureImage() }
+        galleryButton.setOnClickListener{displayGallery()}
+    }
 
+    override fun displayGallery() {
+        val intent = Intent(Intent.ACTION_PICK, galleryFileLocation)
+        startActivityForResult(intent, RESULT_GET_IMAGE)
+    }
+
+    override fun onActivityResult(requestCode: Int, resultCode: Int, intent: Intent?) {
+        super.onActivityResult(requestCode, resultCode, intent)
+        if (requestCode == RESULT_GET_IMAGE && resultCode == Activity.RESULT_OK) {
+            intent?.data?.let { presenter.getImageFromImageFileLocation(it) }
+        }
     }
 
     override fun onStart() {
