+++ app/src/main/java/hypr/a255bits/com/hypr/ModelFragmnt/ModelFragmentMVP.kt
@@ -6,20 +6,16 @@ import android.net.Uri
 
 interface ModelFragmentMVP{
     interface view{
-        fun displayGallery()
         fun displayFocusedImage(imageFromGallery: Bitmap)
         fun  showError(errorMesssage: String)
     }
     interface presenter{
-        fun displayGallery()
-        fun  getImageFromImageFileLocation(imageLocation: Uri)
         fun findFacesInImage(imageWithFaces: Bitmap, context: Context)
         fun disconnectFaceDetector()
         fun saveImageDisplayedToPhone()
         fun transformImage(image: Bitmap?)
     }
     interface interactor{
-        fun  uriToBitmap(imageLocation: Uri): Bitmap
         fun  getFacesFromBitmap(imageWithFaces: Bitmap, width: Int, height: Int, context: Context): MutableList<Bitmap>
     }
 }
